Modularised tests.
