+++ test/smoke/src/mocha-runner.js
@@ -5,6 +5,7 @@
 
 var Mocha = require('mocha');
 var path = require('path');
+var fs = require('fs');
 
 var mocha = new Mocha({
 	timeout: 360000,
